pyproject.toml
[tool.poetry]
name = "responsegen"
version = "0.1.0"
description = ""
authors = ["Julian Prester <julian@julianprester.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
responsegen = "responsegen.cli:main"

[tool.poetry.dependencies]
python = "^3.10"
python-docx = "^0.8.11"
pymupdf = "^1.21.1"
click = "^8.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
